little bit more of decoupling
